+++ src/main/java/tterrag/supermassivetech/handlers/HelmetOverlayHandler.java
@@ -27,7 +27,7 @@ public class HelmetOverlayHandler
     private static final ResourceLocation compass = new ResourceLocation(Reference.MOD_TEXTUREPATH, "textures/gui/overlay/compass.png");
     public static List<String> textToRender = new ArrayList<String>();
     public int maxTime = 400, time = maxTime;
-    
+
     @SubscribeEvent
     public void onClientOverlay(RenderGameOverlayEvent.Text event)
     {
@@ -40,25 +40,27 @@ public class HelmetOverlayHandler
             // HUD is off
             if (Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.OFF))
                 return;
-            
+
             int width = event.resolution.getScaledWidth();
             int height = event.resolution.getScaledHeight();
 
             GL11.glEnable(GL11.GL_BLEND);
             OpenGlHelper.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA, GL11.GL_ONE, GL11.GL_ZERO);
             mc.getTextureManager().bindTexture(compass);
-            int v = getCompassAngle(player);            
-            
+            int v = getCompassAngle(player);
+
             GL11.glColor3f(1f, 1f, 1f);
 
-            if (Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.ON) || Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.COMPASS_ONLY))
+            if (Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.ON)
+                    || Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.COMPASS_ONLY))
             {
                 mc.ingameGUI.drawTexturedModalRect((width - 140) / 2 + getZOffset(mc), 2 + getXOffset(mc), v - 10, 256, 140, 16);
                 renderWaypoints(mc, width, player, player.posX, player.posY, player.posZ);
                 mc.ingameGUI.drawTexturedModalRect((width / 2) + getZOffset(mc), 8 + getXOffset(mc), 6, 16, 3, 9);
             }
-            
-            if (Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.ON) || Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.TEXT_ONLY))
+
+            if (Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.ON)
+                    || Utils.doStatesMatch(player, PowerUps.HUD, 3, GravityArmorHandler.TEXT_ONLY))
             {
                 renderOverlayText(mc, height, width);
             }
@@ -66,9 +68,9 @@ public class HelmetOverlayHandler
             {
                 textToRender.clear();
             }
-            
+
             updateTime();
-            
+
             GL11.glDisable(GL11.GL_ALPHA_TEST);
 
         }
@@ -78,29 +80,29 @@ public class HelmetOverlayHandler
             time = maxTime;
         }
     }
-    
+
     private void renderWaypoints(Minecraft mc, int width, EntityPlayer player, double x, double y, double z)
     {
         for (Waypoint wp : Waypoint.waypoints)
         {
             if (!wp.players.contains(player.getCommandSenderName()))
                 continue;
-            
-            double w = (double) width;
-            
+
+            double w = width;
+
             double angle = angleTo(player, wp.x + 0.5, wp.y + 0.5, wp.z + 0.5);
-            angle *= (w - 16) / (double) 360;
+            angle *= (w - 16) / 360;
             angle += w / 2;
-            
+
             Color c = wp.getColor();
-            
+
             GL11.glColor3f((float) c.getRed() / 255, (float) c.getGreen() / 255, (float) c.getBlue() / 255);
-            
+
             int normal = normalizeAngle(w, angle);
-            
+
             mc.ingameGUI.drawTexturedModalRect(normal + getZOffset(mc), 10 + getXOffset(mc), 0, 16, 5, 16);
-            
-            if (normal < ((width / 2 + 4) + getZOffset(mc)) && normal > ((width / 2 - 4) + getZOffset(mc))) 
+
+            if (normal < ((width / 2 + 4) + getZOffset(mc)) && normal > ((width / 2 - 4) + getZOffset(mc)))
             {
                 mc.ingameGUI.drawCenteredString(mc.fontRenderer, wp.getName(), width / 2, 20, 0xFFFFFF);
                 mc.getTextureManager().bindTexture(compass);
@@ -108,7 +110,6 @@ public class HelmetOverlayHandler
         }
         GL11.glColor3f(1f, 1f, 1f);
     }
-    
 
     private void renderOverlayText(Minecraft mc, int height, int width)
     {
@@ -117,15 +118,15 @@ public class HelmetOverlayHandler
             mc.ingameGUI.drawString(mc.fontRenderer, textToRender.get(i), 5, height - 10 * (i + 1), 0xFFFFFF);
         }
     }
-    
+
     private void updateTime()
     {
-        if (textToRender.isEmpty()) 
+        if (textToRender.isEmpty())
         {
             time = maxTime;
             return;
         }
-        
+
         if (time > 0)
         {
             time -= textToRender.size();
@@ -141,9 +142,9 @@ public class HelmetOverlayHandler
     {
         double dx = player.posX - x;
         double dz = player.posZ - z;
-             
+
         double angleRaw = player.rotationYawHead + Math.toDegrees(Math.atan(dx / dz));
-        
+
         if (dx < 0 & dz < 0)
         {
             angleRaw = angleRaw - 180;
@@ -152,10 +153,10 @@ public class HelmetOverlayHandler
         {
             angleRaw = angleRaw + 180;
         }
-                
+
         while (angleRaw < 0)
             angleRaw += 360;
-        
+
         return (180 + angleRaw) % 360;
     }
 
@@ -167,16 +168,28 @@ public class HelmetOverlayHandler
         return yaw + 3; // arbitrary number to get the texture to line up...can
                         // be changed if texture changes
     }
-    
+
     private int normalizeAngle(double width, double angle)
     {
         double normal = width - angle;
         if (normal < 0)
         {
-            normal = (Math.abs(normal) + (Math.abs(normal) * (34d / width))); // I don't know why 34 don't ask me, probably int -> double inaccuracy
+            normal = (Math.abs(normal) + (Math.abs(normal) * (34d / width))); // I
+                                                                              // don't
+                                                                              // know
+                                                                              // why
+                                                                              // 34
+                                                                              // don't
+                                                                              // ask
+                                                                              // me,
+                                                                              // probably
+                                                                              // int
+                                                                              // ->
+                                                                              // double
+                                                                              // inaccuracy
             normal = width - normal;
         }
-        
+
         if (normal < (width / 2))
         {
             normal = Math.max((width / 2) - 72, normal);
@@ -185,15 +198,15 @@ public class HelmetOverlayHandler
         {
             normal = Math.min((width / 2) + 68, normal);
         }
-        
+
         return (int) normal;
     }
-    
+
     private int getXOffset(Minecraft mc)
     {
         return BossStatus.statusBarTime > 0 || BossStatus.bossName != null ? 18 : mc.gameSettings.showDebugInfo ? 23 : 0;
     }
-    
+
     private int getZOffset(Minecraft mc)
     {
         return mc.gameSettings.showDebugProfilerChart ? -35 : 0;
